+++ lib/parsers.c
@@ -961,7 +961,7 @@ ping_drop:
 
 			eff_buf.token = &wsi->u.ws.rx_user_buffer[
 						   LWS_SEND_BUFFER_PRE_PADDING];
-			eff_buf.token_len = wsi->u.ws.rx_user_buffer_head;
+			eff_buf.token_len = (int)wsi->u.ws.rx_user_buffer_head;
 
 			if (lws_ext_callback_for_each_active(wsi,
 				LWS_EXT_CALLBACK_EXTENDED_PAYLOAD_RX,
@@ -981,7 +981,7 @@ ping_drop:
 
 		eff_buf.token = &wsi->u.ws.rx_user_buffer[
 						LWS_SEND_BUFFER_PRE_PADDING];
-		eff_buf.token_len = wsi->u.ws.rx_user_buffer_head;
+		eff_buf.token_len = (int)wsi->u.ws.rx_user_buffer_head;
 		
 		if (lws_ext_callback_for_each_active(wsi,
 				LWS_EXT_CALLBACK_PAYLOAD_RX, &eff_buf, 0) < 0)
